+++ Src/Models/User.cs
@@ -18,7 +18,7 @@ namespace Quizlet_App_Server.Models
         [BsonElement("user_name")] public string UserName { get; set; } = string.Empty;
         [BsonElement("email")] public string Email { get; set; } = string.Empty;
         //[BsonElement("avatar")] public string Avatar { get; set; } = string.Empty;
-        [BsonElement("avatar")] public List<byte> Avatar { get; set; } = new List<byte>();
+        [BsonElement("avatar")] public List<int> Avatar { get; set; } = new List<int>();
         [BsonElement("date_of_birth")] public string DateOfBirth { get; set; } = "1999-01-01";
         [BsonElement("time_created")] public long TimeCreated { get; set; } = TimeHelper.UnixTimeNow;
         [BsonElement("documents")] public Documents Documents { get; set; } = new Documents();
@@ -96,7 +96,7 @@ namespace Quizlet_App_Server.Models
         [BsonElement("user_name")] public string UserName { get; set; } = string.Empty;
         [BsonElement("email")] public string Email { get; set; } = string.Empty;
         //[BsonElement("avatar")] public string Avatar { get; set; } = string.Empty;
-        [BsonElement("avatar")] public List<byte> Avatar { get; set; } = new List<byte>();
+        [BsonElement("avatar")] public List<int> Avatar { get; set; } = new List<int>();
         [BsonElement("date_of_birth")] public string DateOfBirth { get; set; } = "1999-01-01";
         [BsonElement("time_created")] public long TimeCreated { get; set; } = TimeHelper.UnixTimeNow;
         [BsonElement("documents")] public Documents Documents { get; set; } = new Documents();
@@ -123,7 +123,7 @@ namespace Quizlet_App_Server.Models
         [BsonElement("user_name")] public string? UserName { get; set; } = string.Empty;
         [BsonElement("email")] public string? Email { get; set; } = string.Empty;
         //[BsonElement("avatar")] public string Avatar { get; set; } = string.Empty;
-        [BsonElement("avatar")] public List<byte>? Avatar { get; set; } = new List<byte>();
+        [BsonElement("avatar")] public List<int>? Avatar { get; set; } = new List<int>();
         [BsonElement("date_of_birth")] public string? DateOfBirth { get; set; } = string.Empty;
         [BsonElement("setting")] public UserSetting? Setting { get; set; } = new UserSetting();
     }
